QT-56:Fix delete image of none image product

+++ src/main/java/com/spring/app/products/service/ProductService.java
@@ -187,7 +187,8 @@ public class ProductService {
     }
 
     public void processDeleteProductImage(String customerCode, String image) {
-        storageService.setPath(getImageScope(customerCode)).deleteByName(image);
+        if (image != null)
+            storageService.setPath(getImageScope(customerCode)).deleteByName(image);
     }
 
     public String getImageScope(String path) {
